+++ app/javascript/dashboard/App.vue
@@ -28,10 +28,15 @@ export default {
 
   mounted() {
     this.$store.dispatch('setUser');
+    this.setLocale(window.chatwootConfig.selectedLocale);
     this.initializeAccount();
   },
 
   methods: {
+    setLocale(locale) {
+      Vue.config.lang = locale;
+    },
+
     async initializeAccount() {
       const { pathname } = window.location;
       const accountId = accountIdFromPathname(pathname);
@@ -39,7 +44,7 @@ export default {
       if (accountId) {
         await this.$store.dispatch('accounts/get');
         const { locale } = this.getAccount(accountId);
-        Vue.config.lang = locale;
+        this.setLocale(locale);
       }
     },
   },
