+++ app/views/layouts/vueapp.html.erb
@@ -35,7 +35,8 @@
         <% if ENV['VAPID_PUBLIC_KEY'] %>
         vapidPublicKey: new Uint8Array(<%= Base64.urlsafe_decode64(ENV['VAPID_PUBLIC_KEY']).bytes %>),
         <% end %>
-        enabledLanguages: <%= available_locales_with_name.to_json.html_safe %>
+        enabledLanguages: <%= available_locales_with_name.to_json.html_safe %>,
+        selectedLocale: '<%= I18n.locale %>'
       }
       window.globalConfig = <%= raw @global_config.to_json %>
     </script>
